separated the functionality from the signup component and put it on the login page and deconstructed the functions and values from the props on the signup component
