+++ packages/client/package.json
@@ -17,6 +17,7 @@
     "react-icons": "^4.1.0",
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.1",
+    "sass": "^1.41.1",
     "web-vitals": "^0.2.4"
   },
   "scripts": {
