+++ packages/client/src/components/SignUpForm/index.js
@@ -1,30 +1,36 @@
 import React, { useState } from 'react'
 import './SignUpForm.scss';
 import axios from 'axios'
+import bcrypt from 'bcrypt'
 
 
 
-export default function SignUpForm() {
-    const emptySignUpForm = { //the initial state of the signup form
-        username: '',
-        password: '',
-        emailAddress: '',
-    }
+export default function SignUpForm(props) {
+    const { handleInputChange, handleSignUp, data } = props
     
-    const[data, setData] = useState(emptySignUpForm) //sets the initial state of the signup form
+    //const[data, setData] = useState(emptySignUpForm) //sets the initial state of the signup form
 
-    const handleInputChange = (event) => { //updates the data of the signup form
-        console.log(event.target.value)
-        setData({
-            ...data, //keeps the previous values of the form
-            [event.target.name]: event.target.value, //changes the selected form field
-        })
-    }
-    const handleSignUp = async (event) => { //the function that runs when the form is submitted
-        console.log(data)
-        event.preventDefault()
+    // const handleInputChange = (event) => { //updates the data of the signup form
+    //     console.log(event.target.value)
+    //     setData({
+    //         ...data, //keeps the previous values of the form
+    //         [event.target.name]: event.target.value, //changes the selected form field
+    //     })
+    // }
+    // const handleSignUp = async (event) => { //the function that runs when the form is submitted
+    //     console.log(data)
+    //     event.preventDefault()
         
-    }
+    //     try {
+    //         await axios.post('auth/signup', {
+    //             username: data.username,
+    //             passwordHash: data.password,
+    //             email: data.emailAddress,
+    //         })
+    //     } catch (error) {
+    //         console.log(error)
+    //     }
+    // }
 
     return (
         <div>
